+++ storage/storage.go
@@ -14,4 +14,7 @@ type Storage interface {
 
 	CreateList(list *model.List) error
 	GetOwnerLists(ownerID string) ([]*model.List, error)
+	GetList(id string) (*model.List, error)
+	AddOFFProductToList(listID string, offProduct *model.OFFProduct) error
+	GetOFFProducts(listID string) ([]*model.OFFProduct, error)
 }